+++ perf/perf_common.h
@@ -21,7 +21,7 @@ std::string readFile(const std::string& filename);
 // 多次运行测试并计算平均时间（微秒）
 long long measurePerformance(const std::string& name,
                              std::function<void()> test_func,
-                             int iterations = 1000);
+                             size_t iterations = 1000);
 
 
 // 相对性能测试函数
@@ -29,7 +29,7 @@ bool relativePerformance(const std::string& test_name,
                         std::function<void()> test_func,
                         const std::string& base_name,
                         std::function<void()> base_func,
-                        int iterations = 100,
+                        size_t iterations = 100,
                         long long min_time_ms = 0,
                         double overhead_percent = 0.0);
